requirements.txt
# Для работы с Telegram-ботом, Todoist и Yougile
python-telegram-bot==22.1
requests==2.31.0

# Для поддержки yougile_api
pydantic==1.10.14

# Для работы с YandexGPT
requests
